Move BalanceTable test to its own file
